Add an option to check if the active window is fullscreen
This commit adds a CHECK_FULLSCREEN option and a function to get the
active window and check if the EWMH fullscreen state property is set.

If the bar is hidden and CHECK_FULLSCREEN is 1 and the active window is
fullscreen then the loop will just sleep and not show the bar(s).

Checking for hidden is necessary so the bar can still hide when the
mouse is in the region to show the bar and a window is made fullscreen
or another desktop is switched with a fullscreen window and then the mouse
leaves the region.